chore: rename metadata package
