+++ internal/ecs/metadata/metadata.go
@@ -1,4 +1,4 @@
-package v3
+package metadata
 
 // Copyright 2017-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
