+++ internal/ecs/metadata/metadata_test.go
@@ -1,4 +1,4 @@
-package v3
+package metadata_test
 
 import (
 	"io/ioutil"
@@ -6,28 +6,28 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	"github.com/newrelic/nri-ecs/internal/ecs/metadata"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestClusterARNFromTask(t *testing.T) {
 	taskARN := "arn:aws:ecs:us-west-2:111111111111:task/ecs-local-cluster/37e873f6-37b4-42a7-af47-eac7275c6152"
-	clusterARN := ClusterARNFromTask(taskARN, "ecs-local-cluster")
+	clusterARN := metadata.ClusterARNFromTask(taskARN, "ecs-local-cluster")
 	assert.Equal(t, "arn:aws:ecs:us-west-2:111111111111:cluster/ecs-local-cluster", clusterARN)
 	otherTaskARN := "arn:aws:ecs:eu-west-1:725889879812:task/af32c116-75e0-4f45-aba1-fc2a203ea5d3"
-	otherClusterARN := ClusterARNFromTask(otherTaskARN, "foobar")
+	otherClusterARN := metadata.ClusterARNFromTask(otherTaskARN, "foobar")
 	assert.Equal(t, "arn:aws:ecs:eu-west-1:725889879812:cluster/foobar", otherClusterARN)
 }
 
 func TestResourceNameAndARNBase(t *testing.T) {
 	taskARN := "arn:aws:ecs:us-west-2:111111111111:task/ecs-local-cluster/37e873f6-37b4-42a7-af47-eac7275c6152"
-	resourceName, baseARN := ResourceNameAndARNBase(taskARN)
+	resourceName, baseARN := metadata.ResourceNameAndARNBase(taskARN)
 	assert.Equal(t, "arn:aws:ecs:us-west-2:111111111111", baseARN)
 	assert.Equal(t, "ecs-local-cluster/37e873f6-37b4-42a7-af47-eac7275c6152", resourceName)
-
 }
 
 func TestMetadataResponse(t *testing.T) {
-	var taskResponseJSONFile = "testdata/task_response.json"
+	taskResponseJSONFile := "testdata/task_response.json"
 	taskJSON, err := ioutil.ReadFile(taskResponseJSONFile)
 	assert.NoError(t, err)
 
@@ -38,7 +38,7 @@ func TestMetadataResponse(t *testing.T) {
 	defer testServer.Close()
 
 	client := &http.Client{}
-	response, err := MetadataResponse(client, testServer.URL)
+	response, err := metadata.MetadataResponse(client, testServer.URL)
 	assert.NoError(t, err)
 
 	assert.Equal(t, taskJSON, response)
