+++ pkg/ecs/entity.go
@@ -4,7 +4,7 @@ import (
 	"github.com/newrelic/infra-integrations-sdk/data/metric"
 	"github.com/newrelic/infra-integrations-sdk/integration"
 
-	v3 "github.com/newrelic/nri-ecs/internal/metadata/v3"
+	"github.com/newrelic/nri-ecs/internal/ecs/metadata"
 )
 
 const (
@@ -13,7 +13,7 @@ const (
 )
 
 func NewClusterEntity(clusterARN string, i *integration.Integration) (*integration.Entity, error) {
-	clusterName, arnPrefix := v3.ResourceNameAndARNBase(clusterARN)
+	clusterName, arnPrefix := metadata.ResourceNameAndARNBase(clusterARN)
 	clusterEntity, err := i.Entity("cluster/"+clusterName, arnPrefix)
 	if err != nil {
 		return nil, err
@@ -24,7 +24,6 @@ func NewClusterEntity(clusterARN string, i *integration.Integration) (*integrati
 // AddClusterInventoryLocalEntity adds some ecs attributes as inventory
 // to the integration's local entity.
 func AddClusterInventoryToLocalEntity(clusterName, clusterARN, awsRegion string, launchType LaunchType, integration *integration.Integration) error {
-
 	entity := integration.LocalEntity()
 	err := entity.SetInventoryItem("host", "ecsClusterName", clusterName)
 	if err != nil {
